% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_SQL_duplicates.R
\name{check_SQL_duplicates}
\alias{check_SQL_duplicates}
\title{Checks database for duplicate entries}
\usage{
check_SQL_duplicates(con, dm_obj)
}
\arguments{
\item{con}{a database connection object}

\item{dm_obj}{a dm object}
}
\description{
Checks database for duplicate entries
}
