% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{f1db_cleardb}
\alias{f1db_cleardb}
\title{Deletes database files of the exist, for internal use only by f1db_connect}
\usage{
f1db_cleardb()
}
\description{
Deletes database files of the exist, for internal use only by f1db_connect
}
