% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1db_collect_tables.R
\name{f1db_collect_tables}
\alias{f1db_collect_tables}
\title{Save all f1db tables as global variable tbl's with the same name}
\usage{
f1db_collect_tables(con)
}
\arguments{
\item{con}{A duckdb database connection object}
}
\value{
Saves all 13 Formula One tables as global objects
}
\description{
Save all f1db tables as global variable tbl's with the same name
}
\examples{
\dontrun{f1db_collect_tables(con)}
}
