% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1db_connect.R
\name{f1db_connect}
\alias{f1db_connect}
\title{Downloads configures connect to database}
\usage{
f1db_connect(collect_tables = FALSE)
}
\arguments{
\item{collect_tables}{Boolean Value for additionally executing the f1db_collect_tables(con) function}
}
\value{
A SQLite database connection object
}
\description{
Downloads configures connect to database
}
\examples{
\dontrun{con <- f1db_connect()}

}
