% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{f1db_disconnect}
\alias{f1db_disconnect}
\title{Disconnects from database}
\usage{
f1db_disconnect(con, shutdown = TRUE)
}
\arguments{
\item{con}{A SQLite database connection object}

\item{shutdown}{Boolean value for additionally shutting down the database connection}
}
\description{
Disconnects from database
}
