% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1db_draw.R
\name{f1db_draw}
\alias{f1db_draw}
\title{Constructs database schema diagram.}
\usage{
f1db_draw(con)
}
\arguments{
\item{con}{A duckdb database connection object}
}
\value{
A duckdb database connection object.
}
\description{
Constructs database schema diagram.
}
