% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{f1db_get_dm}
\alias{f1db_get_dm}
\title{Returns dm object for f1db database, for internal use only by f1db_connect}
\usage{
f1db_get_dm()
}
\value{
dm_obj
}
\description{
Returns dm object for f1db database, for internal use only by f1db_connect
}
