% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1db_schema_info.R
\name{f1db_schema_info}
\alias{f1db_schema_info}
\title{Outputs tbl that gives descriptive information about the tables and their contents}
\usage{
f1db_schema_info(con)
}
\arguments{
\item{con}{A sqlite database connection object}
}
\value{
A tbl of database info
}
\description{
Outputs tbl that gives descriptive information about the tables and their contents
}
