% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_load_packages.R
\name{install_load_packages}
\alias{install_load_packages}
\title{Installs all dependencies, for internal use only}
\usage{
install_load_packages(package1, ...)
}
\arguments{
\item{package1}{the first package listed that needs to be installed}

\item{...}{the remaining n packages}
}
\description{
Installs all dependencies, for internal use only
}
