% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1db_schema_info.R
\name{prop_function}
\alias{prop_function}
\title{Helper function used inside schema_info}
\usage{
prop_function(t)
}
\arguments{
\item{t}{value indicating dimensions}
}
\value{
dimension string
}
\description{
Helper function used inside schema_info
}
